requirements.txt
django
python-telegram-bot
Pillow
pre-commit